Core: rename MultipleSubscriptionSnapshotGenerator to SubscriptionMultipleSnapshotGenerator
